requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
dateparser==1.2.0
distlib==0.3.8
filelock==3.13.1
frozenlist==1.4.1
identify==2.5.33
idna==3.6
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.0.5
nodeenv==1.8.0
platformdirs==4.2.0
pre-commit==3.6.0
pycodestyle==2.11.1
pycryptodome==3.20.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyflakes==3.2.0
python-binance==1.0.19
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
ruff==0.2.1
six==1.16.0
typing_extensions==4.9.0
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.0
virtualenv==20.25.0
websockets==12.0
yarl==1.9.4
